add delete student and related action

+++ admin/controllers/student.php
@@ -2,6 +2,10 @@
 
 require '../models/Database.php';
 require '../models/Student.php';
+require '../models/Performance.php';
+require '../models/Business.php';
+require '../models/SchoolApplication.php';
+require '../models/Employee.php';
 
 header('Content-Type: application/json');
 
@@ -35,6 +39,23 @@ if ($_GET['action'] == 'upsertStudent') {
 }
 
 if ($_GET['action'] == 'deleteStudent') {
-    $sRepo->delete($_POST['id']);
-    echo json_encode('ok');
+    $eRepo = new Employee(Database::dbConnect());
+    session_start();
+    $email = $_SESSION['email'];
+    $login = $eRepo->login($email, $_POST['password']);
+    if ($login) {
+        $pRepo = new Performance(Database::dbConnect());
+        $pRepo->removeByStudent($_POST['id']);
+        
+        $bRepo = new Business(Database::dbConnect());
+        $bRepo->removeByStudent($_POST['id']);
+        
+        $saRepo = new SchoolApplication(Database::dbConnect());
+        $saRepo->removeByStudent($_POST['id']);
+        
+        $sRepo->delete($_POST['id']);
+        echo json_encode('ok');
+    } else {
+        echo json_encode('fail');
+    }
 }
